+++ compose-app/src/main/java/com/mapbox/maps/compose/testapp/examples/style/StyleCompositionActivity.kt
@@ -12,7 +12,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.FloatingActionButton
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -43,23 +42,15 @@ import com.mapbox.maps.extension.compose.style.layers.generated.TextColor
 import com.mapbox.maps.extension.compose.style.layers.generated.TextField
 import com.mapbox.maps.extension.compose.style.layers.generated.TextSize
 import com.mapbox.maps.extension.compose.style.layers.generated.Transition
-import com.mapbox.maps.extension.compose.style.projection.Projection
-import com.mapbox.maps.extension.compose.style.projection.Projection.Companion.GLOBE
-import com.mapbox.maps.extension.compose.style.projection.Projection.Companion.MERCATOR
-import com.mapbox.maps.extension.compose.style.projection.Projection.Companion.default
 import com.mapbox.maps.extension.compose.style.sources.generated.GeoJSONData
 import com.mapbox.maps.extension.compose.style.sources.generated.rememberGeoJsonSourceState
 import com.mapbox.maps.extension.style.expressions.generated.Expression
-import com.mapbox.maps.extension.style.sources.generated.geoJsonSource
 
 /**
  * Example to showcase usage of runtime styling with compose.
  */
 @OptIn(MapboxExperimental::class)
 public class StyleCompositionActivity : ComponentActivity() {
-  // Standard style default projection is GLOBE so we make sure that MERCATOR is before and after
-  // default so the map changes visually.
-  private val projections = listOf(default, MERCATOR, GLOBE, MERCATOR)
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
@@ -78,10 +69,6 @@ public class StyleCompositionActivity : ComponentActivity() {
 
       var centerLocation by rememberSaveable {
         mutableStateOf(CityLocations.HELSINKI)
-      }
-
-      var currentProjectionIdx by remember {
-        mutableStateOf(0)
       }
 
       var textColor by remember {
@@ -154,24 +141,6 @@ public class StyleCompositionActivity : ComponentActivity() {
               ) {
                 Text(modifier = Modifier.padding(10.dp), text = "Toggle Style")
               }
-              val enableProjectionBt = mapViewportState.cameraState.zoom < 6.0
-              FloatingActionButton(
-                modifier = Modifier.padding(bottom = 10.dp),
-                backgroundColor = if (enableProjectionBt) MaterialTheme.colors.secondary else Color.LightGray,
-                onClick = {
-                  if (enableProjectionBt) {
-                    currentProjectionIdx = (currentProjectionIdx + 1) % projections.size
-                  }
-                },
-                shape = RoundedCornerShape(16.dp),
-              ) {
-                val nextProjectionName =
-                  projections[(currentProjectionIdx + 1) % projections.size].friendlyName()
-                Text(
-                  modifier = Modifier.padding(10.dp),
-                  text = "Change projection to $nextProjectionName"
-                )
-              }
             }
           }
         ) {
@@ -205,7 +174,6 @@ public class StyleCompositionActivity : ComponentActivity() {
                     }
                   }
                 ),
-                projection = projections[currentProjectionIdx]
               )
             }
           ) {
@@ -262,13 +230,5 @@ public class StyleCompositionActivity : ComponentActivity() {
     const val ZOOM: Double = 9.0
     private var count = 0
     private const val OFFSET = 0.03
-  }
-
-  @OptIn(MapboxExperimental::class)
-  private fun Projection.friendlyName(): String = when (this) {
-    GLOBE -> "globe"
-    MERCATOR -> "mercator"
-    default -> "default"
-    else -> "Unknown"
   }
 }